Implement join space mark
