+++ next/components/space.js
@@ -4,18 +4,14 @@ import { useCallback, useEffect, useState } from "react";
 import InternalLink from "./internalLink";
 import {
   no_scroll_bar,
-  shadow_100,
-  shadow_200,
-  makeSquare,
 } from "../styles/globalCss";
-import { h3_36_bold, p_18_semibold, p_16_semibold } from "../styles/textStyles";
-import SpaceLogo from "@/components/spaceLogo";
+import { h3_36_bold, p_16_semibold } from "../styles/textStyles";
 import { useWindowSize } from "../frontedUtils/hooks";
 import { setCookie } from "frontedUtils/cookie";
-import { Button } from "@osn/common-ui";
-import nextApi from "services/nextApi";
+import { loginAddressSelector } from "store/reducers/accountSlice";
+import SpaceListItem from "./spaceListItem";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchJoinedSpace, joinedSpacesSelector, loginAddressSelector } from "store/reducers/accountSlice";
+import { fetchJoinedSpace } from "store/reducers/accountSlice";
 
 const Title = styled.div`
   ${h3_36_bold};
@@ -45,75 +41,6 @@ const ItemsWrapper = styled.div`
     `}
 `;
 
-const Item = styled.div`
-  flex: 0 0 auto;
-  border: 1px solid #f0f3f8;
-  ${shadow_100};
-  background: #ffffff;
-  padding: 24px;
-  cursor: pointer;
-  width: 163.33px;
-
-  :hover {
-    border-color: #e2e8f0;
-    ${shadow_200}
-  }
-`;
-
-const IconWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
-const Icon = styled.div`
-  ${makeSquare(64)};
-  margin-bottom: 16px;
-`;
-
-const Name = styled.div`
-  white-space: nowrap;
-  ${p_18_semibold};
-  color: #2e343d;
-  text-transform: capitalize;
-`;
-
-const Symbol = styled.div`
-  font-size: 14px;
-  line-height: 24px;
-  color: #a1a8b3;
-`;
-
-const Divider = styled.div`
-  min-width: 116px;
-  height: 1px;
-  background: #f0f3f8;
-  margin: 12px 0;
-`;
-
-const ActiveWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  font-size: 14px;
-  line-height: 24px;
-  color: #a1a8b3;
-`;
-
-const ActiveCircle = styled.div`
-  ${makeSquare(6)};
-  border-radius: 50%;
-  background: #56ca2f;
-  margin-right: 8px;
-`;
-
-const Count = styled.span`
-  margin-left: auto;
-`;
-
-const ActiveCount = styled(Count)`
-  color: #506176;
-`;
-
 const TitleWrapper = styled.div`
   display: flex;
   align-items: center;
@@ -135,38 +62,12 @@ const ButtonWrapper = styled.div`
 export default function Space({ spaces, showAllSpace }) {
   const dispatch = useDispatch();
   const address = useSelector(loginAddressSelector);
-  const joinedSpaces = useSelector(joinedSpacesSelector);
-
-  const isSpaceJoined = useCallback(
-    (spaceName) => !!joinedSpaces.find(item => item.space === spaceName),
-    [joinedSpaces]
-  );
 
   useEffect(() => {
     if (!address) {
       return;
     }
     dispatch(fetchJoinedSpace(address))
-  }, [address]);
-
-  const joinSpace = useCallback(async (spaceName) => {
-    if (!address) {
-      return;
-    }
-    const { result } = await nextApi.post(`account/${address}/joined-spaces`, { space: spaceName });
-    if (result) {
-      dispatch(fetchJoinedSpace(address));
-    }
-  }, [address]);
-
-  const leaveSpace = useCallback(async (spaceName) => {
-    if (!address) {
-      return;
-    }
-    const { result } = await nextApi.delete(`account/${address}/joined-spaces/${spaceName}`);
-    if (result) {
-      dispatch(fetchJoinedSpace(address));
-    }
   }, [address]);
 
   const [show, setShow] = useState(showAllSpace === "1");
@@ -207,31 +108,7 @@ export default function Space({ spaces, showAllSpace }) {
         {(show ? sortedSpaces : sortedSpaces.slice(0, showCount)).map(
           ([name, space], index) => (
             <InternalLink href={`/space/${name}`} key={index}>
-              <Item>
-                <IconWrapper>
-                  <Icon>
-                    <SpaceLogo spaceId={name} />
-                  </Icon>
-                  <Name>{space.name}</Name>
-                  <Symbol>{space.symbol ?? "-"}</Symbol>
-                </IconWrapper>
-                <Divider />
-                <ActiveWrapper>
-                  <ActiveCircle />
-                  <InternalLink href={`/space/${name}?tab=active`}>
-                    Active
-                  </InternalLink>
-                  <Count>
-                    <ActiveCount>{space.activeProposalsCount ?? 0}</ActiveCount>
-                    /{space.proposalsCount}
-                  </Count>
-                </ActiveWrapper>
-                {!isSpaceJoined(name) ? (
-                  <Button onClick={() => joinSpace(name)}>Join</Button>
-                ) : (
-                  <Button onClick={() => leaveSpace(name)}>Leave</Button>
-                )}
-              </Item>
+              <SpaceListItem name={name} space={space} />
             </InternalLink>
           )
         )}
